% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUC.R
\name{aac}
\alias{aac}
\title{compute area above the curve (1-auc)}
\usage{
aac(object, model.id = NULL, batch = NULL, average = c("curve", "area"),
  treatment.only = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{A model id}

\item{batch}{A list with treatment and control}

\item{average}{For batch. Either "curve" or "area". See details}

\item{treatment.only}{default {TRUE}. If TRUE only treatment periode will be considered}
}
\value{
aac values for model.id or treatment and control aac for batch
}
\description{
Given a model.id or batch it will return
aac (area above the curve)
}
\details{
If average = "curve", aac of average curve will be calculated. If average = "area", aac value for each curve will be calculated seperately and values will be average after.
}
\examples{
data(pdxe)
aac(pdxe, model.id = "X.015.BY19")
# creat a experiment desing
myDesign = list(batch.name="myBatch", treatment=c("X.015.BY19"), control=c("X.015.uned"))
aac(pdxe, batch = myDesign)

}

