% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expDesign_manuplation.R
\name{addExperimentalDesign}
\alias{addExperimentalDesign}
\title{add a new experimental design}
\usage{
addExperimentalDesign(object, treatment, control = NULL, batch.name = NULL,
  replace = FALSE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{treatment}{The \code{model.id} of treatment}

\item{control}{The \code{model.id} of control}

\item{batch.name}{The \code{batch.name} for new batch}

\item{replace}{If TRUE will replace the old batch with new values}
}
\value{
returns \code{Xeva} dataset with new experimental design added
}
\description{
add a new experimental design
}
\examples{
data(pdxe)
}

