% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{batchNames}
\alias{batchNames}
\title{Get all batch names}
\usage{
batchNames(object)
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}
}
\value{
A \code{Vector} with all batch.name
}
\description{
Get all batch.name from a Xeva dataset
}
\examples{
data(pdxe)
batchNames(pdxe)
}

