% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRECIST.R
\name{bmRECIST}
\alias{bmRECIST}
\title{compute mRECIST for a batch
Computes the mRECIST for a batch.}
\usage{
bmRECIST(
  contr.time = NULL,
  contr.volume = NULL,
  treat.time = NULL,
  treat.volume = NULL,
  min.time = 10
)
}
\arguments{
\item{contr.time}{Time vector for control.}

\item{contr.volume}{Volume vector for control.}

\item{treat.time}{Time vector for treatment.}

\item{treat.volume}{Volume vector for treatment.}

\item{min.time}{Minimum time after which tumor volume will be considered.}
}
\value{
Returns batch response object.
}
\description{
compute mRECIST for a batch
Computes the mRECIST for a batch.
}
\examples{
contr.time <- treat.time  <- c(0, 3, 7, 11, 18, 22, 26, 30, 32, 35)
contr.volume<- contr.time * tan(60*pi/180)
treat.volume<- treat.time * tan(15*pi/180)
bmr <- bmRECIST(contr.time, contr.volume, treat.time, treat.volume)
print(bmr)
}
