% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brca}
\alias{brca}
\title{breast cancer dataset from PDXE}
\format{A Xeva object of PDXE breast cancer dataset}
\usage{
data(brca)
}
\description{
breast cancer dataset from PDXE
}
\keyword{datasets}
