% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{calculateAngle}
\alias{calculateAngle}
\title{Calculate angle between control and treatment groups}
\usage{
calculateAngle(object, ExpDesign = NULL, var = "volume",
  treatment.only = TRUE, plot = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{ExpDesign}{A list with batch.name, treatment and control}

\item{var}{Name of the variable, default \code{volume}}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a batch (control and treatment model ids)
it will return angle (in Degree) between
the linear fit of treatment and control group
}
\examples{
data(pdxe)
# creat a experiment desing
ExpDesign = list(batch.name="myBatch", treatment=c("X.015.BY19"), control=c("X.015.uned"))
df = calculateAngle(object=pdxe, ExpDesign, var = "volume", treatment.only=TRUE, plot=TRUE)

}

