% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{computAngelFor1ExpDesign}
\alias{computAngelFor1ExpDesign}
\title{data(lpdx); object=lpdx
batchNames <- batchNames(lpdx)
expDegI  <- expDesign(lpdx, "PHLC111_P7")}
\usage{
computAngelFor1ExpDesign(object, expDegI, var = "volume",
  treatment.only = TRUE, plot = FALSE, log.y = TRUE)
}
\description{
data(lpdx); object=lpdx
batchNames <- batchNames(lpdx)
expDegI  <- expDesign(lpdx, "PHLC111_P7")
}

