% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_graveyard.R
\name{computAngelFor1ExpDesign_old}
\alias{computAngelFor1ExpDesign_old}
\title{## get time vs volume data with standard error
#'
#' Given a treatment and control model ids
#' it will return a data.fram with time vs volume (or any other variable)
#'
#' @examples
#' data(pdxe)
#' # creat a experiment desing
#' ExpDesign = list(batch.name="myBatch", treatment=c("X.010.BG98"), control=c("X.010.uned"))
#' df = getTimeVarData(object=pdxe, ExpDesign, var = "volume", collapse=TRUE)
#' ## if collapse=FALSE it will not calculate standard error
#' df2= getTimeVarData(object=pdxe, ExpDesign, var = "volume", collapse=FALSE)
#'
#' @param object The \code{Xeva} dataset
#' @param ExpDesign A list with batch.name, treatment and control
#' @param var Name of the variable, default \code{volume}
#' @param collapse Default \code{TRUE}. It will summerize all models which belongs to same treatment or control group
#' @return a \code{data.fram} with treatment, control and batch.name
setGeneric(name = "Old_getTimeVarData", def = function(object, ExpDesign, var, collapse) {standardGeneric("Old_getTimeVarData")} )}
\usage{
computAngelFor1ExpDesign_old(object, expDegI, var = "volume",
  treatment.only = TRUE, plot = FALSE)
}

