% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRECIST.R
\name{computemRECIST}
\alias{computemRECIST}
\title{Computes the mRECIST}
\usage{
computemRECIST(best.response, best.average.response)
}
\arguments{
\item{best.response}{Value of best response}

\item{best.average.response}{Value of best average response}
}
\value{
Returns the mRECIST for given volume response
}
\description{
\code{computemRECIST} returns the mRECIST for given volume response
}
\examples{
computemRECIST(best.response=8.722, best.average.response=8.722)
}

