% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoPX_Class.R
\name{creatPharmacoPXSet}
\alias{creatPharmacoPXSet}
\title{if model and drug slot is NULL it will try to infer it
otherwise will use the given model and drug slot}
\usage{
creatPharmacoPXSet(name, molecularProfiles = list(),
  experiment = data.frame(), model = data.frame(), drug = data.frame())
}
\description{
if model and drug slot is NULL it will try to infer it
otherwise will use the given model and drug slot
}

