% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoSet_Class.R
\name{creatXenoSet}
\alias{creatXenoSet}
\title{Creat XenoR class object}
\usage{
creatXenoSet(name, molecularProfiles = list(), experiment = data.frame(),
  model = data.frame(), drug = data.frame())
}
\arguments{
\item{name}{A \code{character} string detailing the name of the dataset}

\item{molecularProfiles}{A \code{list} of ExpressionSet objects containing molecular profiles}

\item{experiment}{A \code{data.frame} containg all experiment information}

\item{model}{A \code{data.frame} containg the annotations for all models used in the experiment}

\item{drug}{A \code{data.frame} containg the annotations for all the drugs profiled in the data set, across all data types}
}
\value{
Returns XenoR object
}
\description{
\code{creatPharmacoPxSet} returns XenoR class object
}
\examples{
geoExp = readRDS("DATA-raw/Geo_Exp.Rda")
pdxe = creatPharmacoPxSet(name = "PDXE",
                          molecularProfiles = list(RNASeq = geoExp$RNASeq),
                          experiment = geoExp$experiment,
                          model = geoExp$model,
                          drug  = geoExp$drug )
save(pdxe, file = "data/pdxe.rda")
data("pdxe")
}

