% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XevaSet_Class.R
\name{creatXevaSet}
\alias{creatXevaSet}
\title{Creat Xeva class object
\code{creatXevaSet} returns Xeva class object}
\usage{
creatXevaSet(name, model = data.frame(), drug = data.frame(),
  experiment = data.frame(), expDesign = list(),
  modelSensitivity = data.frame(), batchSensitivity = data.frame(),
  molecularProfiles = list(), modToBiobaseMap = data.frame())
}
\arguments{
\item{name}{a \code{character} string detailing the name of the dataset}

\item{model}{a \code{data.frame} containg the annotations for all models used in the experiment}

\item{drug}{a \code{data.frame} containg the annotations for all the drugs profiled in the data set, across all data types}

\item{experiment}{a \code{data.frame} containg all experiment information}

\item{molecularProfiles}{a \code{list} of ExpressionSet objects containing different molecular profiles}
}
\value{
Returns Xeva object
}
\description{
Creat Xeva class object
\code{creatXevaSet} returns Xeva class object
}
\examples{
\\code{NULL}
}
