% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoPX_Class.R
\name{documentationExample}
\alias{documentationExample}
\title{Add together two numbers.}
\usage{
documentationExample(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{Whatever}{u put here}
}
\value{
The sum of \code{x} and \code{y} and \code{z} . z is nothing
}
\description{
Add together two numbers.
}
\examples{
documentationExample(1, 1)
This will be also included
}

