% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_drug.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{Set new drug information in a XevaSet
A method to set the drug information slot in a \code{XevaSet} object.}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in.}

\item{value}{A \code{data.frame} with the new drug annotations.}
}
\value{
updated \code{XevaSet}
}
\description{
Set new drug information in a XevaSet
A method to set the drug information slot in a \code{XevaSet} object.
}
\examples{
data(brca)
drugInfo(brca)<- drugInfo(brca)
}
