% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_drug.R
\name{drugInfo}
\alias{drugInfo}
\title{drugInfo Generic
Generic for drugInfo method}
\usage{
drugInfo(object)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the drug annotations
}
\description{
drugInfo Generic
Generic for drugInfo method
}
\examples{
data(pdxe)
drugInfo(pdxe)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
