% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{drugWaterfall}
\alias{drugWaterfall}
\title{Plot drug response waterfall}
\usage{
drugWaterfall(object, drug, value, col = "#cc4c02", group.by = "patient.id",
  tumor.type = NULL, title = "", yname = "Response")
}
\description{
Plot drug response waterfall
}
\examples{
data(pdxe)
drugWaterfall(pdxe, drug="binimetinib", value="best.avg.response_published",
              col="#E69F00", tumor.type = "CRC", title="Binimetinib",
              yname = "Change in tumor volume (\%)")
}

