% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{expDesign}
\alias{expDesign}
\title{Given a batch.name get batch}
\usage{
expDesign(object, batch.name = NULL)
}
\arguments{
\item{object}{\code{XevaSet}}

\item{object}{\code{batch.name}. If NULL will return all batch in the dataset}
}
\value{
A \code{Vector} with all batch.name
}
\description{
Given a batch.name get batch from a Xeva dataset
}
\examples{
data(brca)
expDesign(brca, batch.name = "X-6047.paclitaxel")
}
