% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{expDesignInfo<-}
\alias{expDesignInfo<-}
\title{expDesignInfo<- Generic
Generic for expDesignInfo replace method}
\usage{
expDesignInfo(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}

\item{value}{A \code{list} with the experiment designs}
}
\value{
Updated \code{XevaSet}
}
\description{
expDesignInfo<- Generic
Generic for expDesignInfo replace method
}
\examples{
data(pdxe)
expDesignInfo(pdxe) <- expDesignInfo(pdxe)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
