% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{expDesignInfo}
\alias{expDesignInfo}
\title{expDesignInfo Generic
Generic for expDesignInfo method}
\usage{
expDesignInfo(object)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{list} with the all experiment designs
}
\description{
expDesignInfo Generic
Generic for expDesignInfo method
}
\examples{
data(pdxe)
expDesignInfo(pdxe)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
