% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_treatment_control_functions.R
\name{expType}
\alias{expType}
\title{Get experiment type (treatment or control) for a given model.id}
\usage{
expType(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
returns \code{treatment} or \code{control}
}
\description{
Get experiment type (treatment or control) for a given model.id
}
\examples{
data(pdxe)
# get experiment type for model.id
getTreatment(object=pdxe, model.id="X-6047.21")
}

