% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{fitGP}
\alias{fitGP}
\title{Fit a GP using Python's GPy package.}
\usage{
fitGP(replicates, source)
}
\arguments{
\item{replicates}{A vector of multiple replicates that user would like to fit GP on}

\item{source}{Data source}
}
\value{
results A list of time, predicted mean, and variance of fit GP
}
\description{
\code{fitGP} fit a GP using Python's GPy package
}
\examples{
fitGP(c("PHLC111_P7.701.A1", "PHLC111_P7.703.A3", "PHLC111_P7.706.B1", "PHLC111_P7.708.B3"), lpdx)
fitGP(c("PHLC111_P7.702.A2", "PHLC111_P7.704.A4", "PHLC111_P7.705.A5", "PHLC111_P7.707.B2"), lpdx)
}

