% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_Sensitivity_Signature.R
\name{geneSensitivityPlot}
\alias{geneSensitivityPlot}
\title{geneSensitivityPlot}
\usage{
geneSensitivityPlot(object, mDataType, feature, drug,
  sensitivity.measure = "slope", standardize = c("SD", "rescale", "log",
  "none"))
}
\description{
Given a Xeva object, feture name and drug name it will plot feture values against sensitivity value
}
\details{
Plot a gene expression against sensitivity signatures for a drug
}
\examples{
data(pdxe)
geneSensitivityPlot(object=pdxe, mDataType="RNASeq", feature="A1BG", drug="binimetinib",
sensitivity.measure="slope", standardize="log")

}
