% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Batch_mol_data.R
\name{getBatchAndMolData}
\alias{getBatchAndMolData}
\title{data(brca)
object=brca; mDataType="RNASeq"; drug="BGJ398"; tissue=NULL; unique.model=TRUE
senType="batch"; sensitivity.measure="abc"}
\usage{
getBatchAndMolData(
  object,
  mDataType,
  drug,
  tissue,
  unique.model,
  senType,
  sensitivity.measure
)
}
\description{
data(brca)
object=brca; mDataType="RNASeq"; drug="BGJ398"; tissue=NULL; unique.model=TRUE
senType="batch"; sensitivity.measure="abc"
}
