% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getBatchName}
\alias{getBatchName}
\title{Extract batch.name for a given model.id}
\usage{
getBatchName(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{vector} with treatment model.id
}
\description{
Extract batch.name for a given model.id
}
\examples{
data(pdxe)
# extract treatment for a given model.id
getBatchName(object=pdxe, model.id="X.1655.uned")
getBatchName(object=pdxe, model.id="X.010.fiab")
}

