% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getControls}
\alias{getControls}
\title{Extract controls for a given model.id
If no control found it will return NULL}
\usage{
getControls(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{vector} with control model.id
}
\description{
Extract controls for a given model.id
If no control found it will return NULL
}
\examples{
data(pdxe)
# extract controls for a given model.id
getControls(object=pdxe, model.id="X.1655.LE11.biib")
# if no control found it will return NULL
getControls(object=pdxe, model.id="X.1655.uned")
}

