% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getExpDesignDF}
\alias{getExpDesignDF}
\title{Given a model.id it will return a data.fram of experiemt design}
\usage{
getExpDesignDF(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a model.id it will return a data.fram of experiemt design
}
\examples{
data(pdxe)
# extract controls for a given model.id
getExpDesignDF(object=pdxe, model.id="X.1655.LE11.biib")
getExpDesignDF(object=pdxe, model.id="X-6047.16")
}

