% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{getExperiment}
\alias{getExperiment}
\title{getExperiment Generic
Generic for getExperiment method}
\usage{
getExperiment(object, model.id)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{list} with the all experiment designs
}
\description{
getExperiment Generic
Generic for getExperiment method
}
\examples{
data(pdxe)
getExperiment(pdxe, model.id="X.1655.LE11.biib")
getExperiment(pdxe, model.id=c("X.1655.LE11.biib", "X.1298.pael") )
}

