% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{getExperimentData}
\alias{getExperimentData}
\title{get experiment time series data for a given biobase.id or model.id}
\usage{
getExperimentData(object, id, id.name, exact.match)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{vector} with control model.id
}
\description{
get experiment time series data for a given biobase.id or model.id
}
\examples{
data(pdxe)
# extract controls for a given model.id
getExperimentData(object=pdxe, id ="X.007.trab", id.type="model.id")
getExperimentData(object=pdxe, id ="X-007", id.type="biobase.id")
}

