% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{getExperimentDF}
\alias{getExperimentDF}
\title{Given a PDX object this will plot drug response curve
i.e. Time vs Volume}
\usage{
getExperimentDF(expSlot, modelID, value)
}
\arguments{
\item{PDX}{a PDX object}
}
\value{
ggplot object
}
\description{
Given a PDX object this will plot drug response curve
i.e. Time vs Volume
}
\examples{
plotDrugResponse(PDX)
}

