% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{getExperimentIds}
\alias{getExperimentIds}
\title{getExperimentIds Generic
Generic for getExperimentIds method}
\usage{
getExperimentIds(object, drug = NULL, drug.match.exact = TRUE,
  tumor.type = NULL)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{list} with the all experiment designs
}
\description{
getExperimentIds Generic
Generic for getExperimentIds method
}
\examples{
data(pdxe)
getExperimentIds(pdxe, drug="paclitaxel", drug.match.exact=TRUE, tumor.type="BRCA")
}

