% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{getGPStatistics}
\alias{getGPStatistics}
\title{Get KL divergence of two GPs}
\usage{
getGPStatistics(controlReplicates, caseReplicates, source)
}
\arguments{
\item{controlReplicates}{Vector of control replicates}

\item{caseReplicates}{Vector of case replicates}

\item{source}{Data source of replicates}
}
\value{
Returns the KL divergence value for two GPs.
}
\description{
\code{getGPStatistics} Get the KL divergence of two GPs.
}
\examples{
getStatistics(c("PHLC111_P7.701.A1", "PHLC111_P7.703.A3"), c("PHLC111_P7.705.A5", "PHLC111_P7.707.B2"), lpdx)

}

