% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{getModelIds}
\alias{getModelIds}
\title{getModelIds Generic
Generic for getModelIds method}
\usage{
getModelIds(object, drug = NULL, drug.match.exact = TRUE,
  tumor.type = NULL)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{list} with the all experiment designs
}
\description{
getModelIds Generic
Generic for getModelIds method
}
\examples{
data(pdxe)
getModelIds(pdxe, drug="paclitaxel", drug.match.exact=TRUE, tumor.type="BRCA")
}

