% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_expressionData_Functions.R
\name{getMolecularProfileAssay}
\alias{getMolecularProfileAssay}
\title{Get a specific assay (SummarizedExperiment) from a XevaSet object's molecularProfiles}
\usage{
getMolecularProfileAssay(object, data.type)
}
\arguments{
\item{object}{The \code{XevaSet} object.}

\item{data.type}{\code{character} name of the assay (e.g. "RNASeq").}
}
\value{
A \code{SummarizedExperiment} corresponding to the given \code{data.type}.
}
\description{
This function provides direct access to a specific assay stored in the
\code{MultiAssayExperiment} slot of a \code{XevaSet} object.
}
\examples{
data(brca)
rna <- getMolecularProfileAssay(brca, "RNASeq")
}
