% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRECIST.R
\name{getmRECIST}
\alias{getmRECIST}
\title{getmRECIST Generic
Generic for getmRECIST method}
\usage{
getmRECIST(object, group.by = "biobase.id")
}
\arguments{
\item{object}{The \code{XenoSet} to retrieve mRECIST from}

\item{group.by}{The name of column which will be mapped to model.id}
}
\value{
a \code{data.frame} with the mRECIST values, rows are drugs and columns are model.id
}
\description{
getmRECIST Generic
Generic for getmRECIST method
}
\examples{
data(pdxe)
# calculate mRECIST for each experiment
setmRECIST(pdxe)<- setmRECIST(pdxe)
getmRECIST(pdxe, group.by="biobase.id")
}

