% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_batch_data.R
\name{getTimeVarData}
\alias{getTimeVarData}
\title{Get time vs volume data with standard error}
\usage{
getTimeVarData(object, batchName = NULL, expDig = NULL, var = "volume",
  treatment.only = FALSE, drug.name = FALSE, vol.normal = FALSE,
  impute.value = TRUE, max.time = NULL)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{var}{Name of the variable, default \code{volume}}

\item{drug.name}{\code{FALSE}. If \code{TRUE} will return drug name also}

\item{ExpDesign}{A list with batch.name, treatment and control}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a batch (treatment and control model ids)
it will return a data.fram with time vs volume (or any other variable)
with standard error calculated. Note that this function do not check if
model.id in given batch belongs to same patient
}
\examples{
data(pdxe)
# creat a experiment desing
ExpDesign = list(batch.name="myBatch", treatment=c("X.010.BG98", "X.010.BG98"), control=c("X.010.uned"))
df = getTimeVarData(object=pdxe, ExpDesign, var = "volume")

}
