% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getTreatment}
\alias{getTreatment}
\title{Get treatment for a given model.id}
\usage{
getTreatment(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{vector} with treatment model.id
}
\description{
Get treatment for a given model.id.
If no treatment found it will return NULL
}
\examples{
data(pdxe)
# extract treatment model.id for a given model.id
getTreatment(object=pdxe, model.id="X.1655.uned")
}
