% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{linear mixed model}
\usage{
lmm(data)
}
\arguments{
\item{data}{a data.frame containg a batch data}
}
\value{
Returns a fit object
}
\description{
Comput the linear mixed model (lmm) statistics for a PDX batch
}
\details{
The input data.frame (data) must contain these columns: model.id, volume, time, exp.type
}
\examples{
data(repdx)
data <- getExperiment(repdx, batch = "P1")$model
lmm(data)

}
