% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{mapModelSlotIds}
\alias{mapModelSlotIds}
\title{map one id type to another in model slot
model.id given a "biobase.id" or "patient id" or any other id that is a column in model slot}
\usage{
mapModelSlotIds(object, id, id.name, map.to)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{id}{The \code{id}}

\item{id.name}{The \code{id} name}

\item{map.to}{The name of the mapped id}
}
\value{
a \code{data.fram} with id and mapped id
}
\description{
map one id type to another in model slot
model.id given a "biobase.id" or "patient id" or any other id that is a column in model slot
}
\examples{
data(pdxe)
mapModelSlotIds(object=pdxe, id="X-007", id.name="biobase.id", map.to="model.id")
}

