% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_modToBiobase.R
\name{model2BiobaseIdMap}
\alias{model2BiobaseIdMap}
\title{model2BiobaseIdMap
Gives model.id to biobase.id mapping datafram}
\usage{
model2BiobaseIdMap(object, mDataType = NULL)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{mDataType}{Data type for which ids to be retrive. Default \code{NULL} will return full datafram}
}
\value{
a \code{data.frame} with the model.id and biobase.id
}
\description{
model2BiobaseIdMap
Gives model.id to biobase.id mapping datafram
}
\examples{
data(brca)
idMap <- model2BiobaseIdMap(brca, mDataType="RNASeq")
head(idMap)
}
\keyword{internal}
