% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{ModelInfo<-}
\alias{ModelInfo<-}
\title{ModelInfo<- Generic
Generic for ModelInfo replace method}
\usage{
ModelInfo(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}

\item{value}{A \code{data.frame} with the new model annotations}
}
\value{
Updated \code{XevaSet}
}
\description{
ModelInfo<- Generic
Generic for ModelInfo replace method
}
\examples{
data(pdxe)
ModelInfo(pdxe) <- ModelInfo(pdxe)
}

