% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{ModelInfo}
\alias{ModelInfo}
\title{modelInfo Generic
Generic for ModelInfo method}
\usage{
ModelInfo(object)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the model annotations
}
\description{
modelInfo Generic
Generic for ModelInfo method
}
\examples{
data(pdxe)
ModelInfo(pdxe)
}

