% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{multiplyBy3}
\alias{multiplyBy3}
\title{This is an example of an internal function called \code{multiplyBy3()}}
\usage{
multiplyBy3(base_number)
}
\arguments{
\item{base_num}{The number to multiply by three}
}
\value{
Returns a numeric vector
}
\description{
Sometimes you want internal functions as part of an R Package built with
RStudio and roxygen2, but you don't want .Rd files created for them
or to have them be visible in the help document following the build process
}
\keyword{internal}
