% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_graveyard.R
\name{Old_getTimeVarData}
\alias{Old_getTimeVarData}
\title{Given a treatment and control model ids
it will return a data.fram with time vs volume (or any other variable)}
\usage{
Old_getTimeVarData(object, ExpDesign, var, collapse)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{ExpDesign}{A list with batch.name, treatment and control}

\item{var}{Name of the variable, default \code{volume}}

\item{collapse}{Default \code{TRUE}. It will summerize all models which belongs to same treatment or control group}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a treatment and control model ids
it will return a data.fram with time vs volume (or any other variable)
}
\examples{
data(pdxe)
# creat a experiment desing
ExpDesign = list(batch.name="myBatch", treatment=c("X.010.BG98"), control=c("X.010.uned"))
df = getTimeVarData(object=pdxe, ExpDesign, var = "volume", collapse=TRUE)
## if collapse=FALSE it will not calculate standard error
df2= getTimeVarData(object=pdxe, ExpDesign, var = "volume", collapse=FALSE)

}

