% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{pasteWithoutNA}
\alias{pasteWithoutNA}
\title{paste a vector elements togather while removing NA}
\usage{
pasteWithoutNA(L, collapse = " + ")
}
\arguments{
\item{L}{A vector with values and NA}

\item{collapse}{Collapse string default " + "}
}
\value{
Returns an string with vector values paste togather
}
\description{
\code{pasteWithoutNA} paste a vector elements togather while removing NA
}
\examples{
L = c("A", NA, "B", NA, NA, "C")
pasteWithoutNA(L, collapse = " + ")
}
