% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdxe}
\alias{pdxe}
\title{Example dataset with 1x1x1 experiment design}
\format{A Xeva object with 1x1x1 experiment design and moleculer data}
\usage{
data(pdxe)
}
\description{
This is PDXE dataset without microarray data.
}
\keyword{datasets}
