% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PDXMI}
\alias{PDXMI}
\title{PDX-MI data}
\format{
An object of class \code{data.frame} with 45 rows and 4 columns.
}
\source{
\url{http://cancerres.aacrjournals.org/lookup/doi/10.1158/0008-5472.CAN-17-0582}
}
\usage{
data(PDXMI)
}
\description{
A dataset containing PDX models minimal information (PDX-MI) standard and corresponding Xeva variable.
}
\details{
For details about PDX-MI, see:

Meehan, Terrence F., et al. "PDX-MI: minimal information for patient-derived tumor xenograft models." Cancer research 77.21 (2017): e62-e66.
}
\keyword{datasets}
