% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRECIST.R
\name{plot.mRECIST}
\alias{plot.mRECIST}
\title{Plot mRECIST for models and drugs}
\usage{
\method{plot}{mRECIST}(df, groupBy = "biobase.id",
  control.name = "untreated")
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
plot
}
\description{
\code{plot.mRECIST} plots the mRECIST
}
\examples{
data(pdxe)
#groupBy = "biobase.id"
control.name = c("untreated")
df = getmRECIST(pdxe)
df = df[1:500,]
plot.mRECIST(df,groupBy = "biobase.id", control.name = "untreated")

}

