% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotBatch}
\alias{plotBatch}
\title{data(lpdx)
plotBatch(lpdx, "PHLC153_P6", treatment.only=FALSE, log.y=TRUE)}
\usage{
plotBatch(object, batchName, treatment.only = FALSE, ...)
}
\description{
data(lpdx)
plotBatch(lpdx, "PHLC153_P6", treatment.only=FALSE, log.y=TRUE)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
