% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotDrugResponse}
\alias{plotDrugResponse}
\title{To plot drug response curve}
\usage{
plotDrugResponse(object, drug, drug.match.exact, tumor.type, control)

plotDrugResponse(object, drug, drug.match.exact, tumor.type, control)
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}

\item{drug}{Name of the drug}

\item{object}{The \code{XevaSet} to replace drug info in}

\item{drug}{Name of the drug}
}
\value{
Updated \code{XevaSet}

Updated \code{XevaSet}
}
\description{
To plot drug response curve
}
\examples{
data(pdxe)
plotDrugResponse(pdxe, drug="LEE011 + binimetinib")
plotDrugResponse(pdxe, drug="paclitaxel", drug.match.exact=TRUE, tumor.type="BRCA")
}

