% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotModelErrorBar}
\alias{plotModelErrorBar}
\title{##df = readRDS("DATA-raw/toPlot_DF.Rda")
##plotModelErrorBar(df)}
\usage{
plotModelErrorBar(df, control.col = "#6baed6", treatment.col = "#fc8d59",
  title = "", xlab = "Time", ylab = "Volume", log.y = TRUE,
  SE.plot = c("errorbar", "ribbon"), aspect.ratio = c(1, NULL))
}
\description{
##df = readRDS("DATA-raw/toPlot_DF.Rda")
##plotModelErrorBar(df)
}
