% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRECIST.R
\name{plotmRECIST}
\alias{plotmRECIST}
\title{Plot mRECIST for models and drugs
\code{plot.mRECIST} plots the mRECIST}
\usage{
plotmRECIST(df, groupBy = "biobase.id", control.name = "untreated")
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
plot
}
\description{
Plot mRECIST for models and drugs
\code{plot.mRECIST} plots the mRECIST
}
\examples{
data(pdxe)
df = getmRECIST(pdxe)
## add tumor.type information
dfMap = mapModelSlotIds(object=pdxe, id=df$model.id, id.name="model.id",
                                map.to="tumor.type", unique=FALSE)
df$tumor.type = dfMap$tumor.type
## select only lung cancer models
lungDf = df[df$tumor.type=="NSCLC",]
plotmRECIST(lungDf, groupBy = "biobase.id", control.name = "untreated")
}

