% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{print.batchResponse}
\alias{print.batchResponse}
\title{Print the batch response}
\usage{
\method{print}{batchResponse}(x, ...)
}
\arguments{
\item{x}{batchResponse object}

\item{...}{Other arguments}
}
\value{
prints the batchResponse
}
\description{
Print the batch response
}
