% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat_expDesign_slot.R
\name{print.pdxBatch}
\alias{print.pdxBatch}
\title{Print the pdx batch}
\usage{
\method{print}{pdxBatch}(x, ...)
}
\arguments{
\item{x}{pdxBatch object}

\item{...}{Other arguments}
}
\description{
Print the pdx batch
}
