% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{print.XevaSet}
\alias{print.XevaSet}
\title{print Xeva object}
\usage{
\method{print}{XevaSet}(object, id = NULL)
}
\arguments{
\item{object}{\code{Xeva} object}

\item{id}{default \code{NULL}, id can be \code{model.id} or \code{batch.name}}
}
\value{
Prints object, model or batch information.
}
\description{
\code{print} displays Xeva object information or model or batch information
}
\examples{
data(pdxe)
# to print object information
print(pdxe)

# to print a model
model.id = modelInfo(pdxe)$model.id[1]
print(pdxe, id = model.id)

# to print a batch
batch.id = batchNames(pdxe)[1]
print(pdxe, id = batch.id)
}

