% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{printAndCapture}
\alias{printAndCapture}
\title{Function to print data.frame in massage}
\usage{
printAndCapture(x)
}
\description{
\code{printAndCapture} prints data.frame in stop or warning functions
}
\examples{
df <- data.frame(a=1:5, b=11:15)
msg <- sprintf("data frame is:\\n\%s", printAndCapture(df))
warning(msg)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
