% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{printf}
\alias{printf}
\title{Function to print}
\usage{
printf(...)
}
\description{
\code{printf} is shortcut for cat(sprintf())
}
\examples{
printf("Hello \%s\\n", "world" )
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
