% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{removeZeroVar}
\alias{removeZeroVar}
\title{Function to remove low variance features}
\usage{
removeZeroVar(df, varCutoff = 0, sort = TRUE)
}
\description{
Function to remove low variance features
}
\examples{
data(cars)
removeZeroVar(cars, varCutoff=0)
}
\keyword{internal}
