% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repdx}
\alias{repdx}
\title{Example PDX dataset}
\format{An object of class \code{XevaSet} of length 1.}
\usage{
data(repdx)
}
\description{
A Xeva object containing anonymize PDX data with replicates. Each batch have 5 replicats.
}
\keyword{datasets}
