% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrixs.R
\name{response}
\alias{response}
\title{compute response}
\usage{
response(object, model.id = NULL, batch = NULL, res.measure = c("angle",
  "mRECIST", "AUC", "angle", "abc"), treatment.only = TRUE, max.time = NULL,
  impute.value = TRUE, min.time = 10, concurrent.time = TRUE,
  vol.normal = F, verbose = TRUE)
}
\arguments{
\item{object}{Xeva object}

\item{model.id}{model id for which response to be computed}

\item{batch}{batch id or experiment design for which response to be computed}

\item{res.measure}{response measure}

\item{treatment.only}{Default \code{FALSE}. If TRUE give data only for non-zero dose periode (if dose data avalible)}

\item{max.time}{maximum time for data}

\item{impute.value}{default \code{FALSE}. If TRUE will impute the values}

\item{min.time}{default \strong{10} days. Used for \emph{mRECIST} computation}

\item{concurrent.time}{default \code{FALSE}. If TRUE will cut the batch data such that control and treatment will end at same time point}

\item{vol.normal}{default \code{TRUE} will use}

\item{verbose}{default \code{TRUE} will print infromation}
}
\value{
returns model or batch response object
}
\description{
\code{response} computes response of a PDX model or batch
}
\examples{
data(brca)
response(brca, model.id="X.1004.BG98", res.measure="mRECIST")

response(brca, batch="X-6047.paclitaxel", res.measure="angle")

ed <- list(batch.name="myBatch", treatment=c("X.6047.LJ16","X.6047.LJ16.trab"),
             control=c("X.6047.uned"))
response(brca, batch=ed, res.measure="angle")

}
