% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{selectModelIds}
\alias{selectModelIds}
\title{To select model ids based on drug name and/or tumor type}
\usage{
selectModelIds(object, drug = NULL, drug.match.exact = TRUE,
  tumor.type = NULL)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{drug}{Name of the \code{drug}}

\item{drug.match.exact}{Default \code{TRUE}}

\item{tumor.type}{Tumor type. Default \code{NULL}}
}
\value{
a \code{vector} with the matched model.ids
}
\description{
To select model ids based on drug name and/or tumor type
}
\examples{
data(pdxe)
selectModelIds(pdxe, drug="paclitaxel", drug.match.exact=TRUE, tumor.type="BRCA")
}

