% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Get sensitivity for an Xeva object}
\usage{
sensitivity(object, type, sensitivity.measure = NULL)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{type}{sensitivity type (either model or batch)}

\item{sensitivity.measure}{Name of the sensitivity.measure. Default \code{NULL}, will return all}
}
\value{
a \code{data.fram} with model or batch id and sensitivity values
}
\description{
Given a Xeva object, it will return sensitivity datafram
}
\examples{
data(cm.pdxe)
head(sensitivity(cm.pdxe, type="batch"))
head(sensitivity(cm.pdxe, type="model"))
}
