% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{setAngle<-}
\alias{setAngle<-}
\title{setAngle<- Generic}
\usage{
setAngle(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}
}
\value{
Updated \code{XevaSet}
}
\description{
Generic for setAngle replace method
}
\examples{
data(pdxe)
setAngle(pdxe) <- setAngle(object = pdxe)
}
