% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{setAngle}
\alias{setAngle}
\title{setAngle generic}
\usage{
setAngle(object, treatment.only = FALSE)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{list} with angle between treatment and control for each batch
}
\description{
Generic for setAngle method
}
\examples{
data(pdxe)
angles <- setAngle(object = pdxe)
}
