% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRECIST.R
\name{setmRECIST}
\alias{setmRECIST}
\title{setmRECIST<- Generic for setmRECIST replace method}
\usage{
setmRECIST(object, min.time = 10)
}
\arguments{
\item{object}{The \code{XevaSet} object}
}
\value{
Updated \code{XevaSet}
}
\description{
setmRECIST<- Generic for setmRECIST replace method
}
\examples{
data(pdxe)
#calculate mRECIST for each experiment
setmRECIST(pdxe)<- setmRECIST(pdxe)
getmRECIST(pdxe)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
