% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrixs.R
\name{setResponse}
\alias{setResponse}
\title{\code{setResponse} sets response of an Xeva object.}
\usage{
setResponse(object, res.measure = c("mRECIST", "slope", "AUC", "angle",
  "abc", "TGI"), min.time = 10, treatment.only = FALSE,
  max.time = NULL, vol.normal = TRUE, impute.value = TRUE,
  concurrent.time = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{Xeva object.}

\item{res.measure}{Response measure; multiple measures are allowed.}

\item{min.time}{Default \strong{10} days. Used for \emph{mRECIST} computation.}

\item{treatment.only}{Default \code{FALSE}. If \code{TRUE}, give data for non-zero dose periods only (if dose data are available).}

\item{max.time}{Maximum time for data.}

\item{vol.normal}{Default \code{TRUE} will use}

\item{impute.value}{Default \code{FALSE}. If \code{TRUE}, impute the missing values.}

\item{concurrent.time}{Default \code{FALSE}. If \code{TRUE}, cut the batch data such that control and treatment will end at same time point.}

\item{verbose}{Default \code{TRUE} will print information}
}
\value{
Returns updated Xeva object.
}
\description{
\code{setResponse} sets response of an Xeva object.
}
\examples{
data(brca)
\dontrun{ brca  <- setResponse(brca, res.measure = c("mRECIST")) }
}
