% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_sensitivity.R
\name{setSensitivity}
\alias{setSensitivity}
\title{add new sensitivity in a Xeva object}
\usage{
setSensitivity(object, type, name, value)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{type}{sensitivity type (either model or batch)}

\item{name}{name of new sensitivity column}

\item{value}{a vector of values. If vector is named, values will be filled by name. Missing values will be NA}
}
\value{
a \code{Xeva} object with updated with updated sensitivity
}
\description{
This will add a add new sensitivity in a Xeva object
}
\examples{
data(pdxe)
s <- sensitivity(pdxe, "model")
pdxe <- setSensitivity(pdxe, "model", "mR", s$mRECIST)
}
