% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{setSlop<-}
\alias{setSlop<-}
\title{setSlop<- Generic}
\usage{
setSlop(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}
}
\value{
Updated \code{XevaSet}
}
\description{
Generic for setSlop replace method
}
\examples{
data(pdxe)
setSlop(pdxe) <- setSlop(object = pdxe)
}

