% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{setSlop}
\alias{setSlop}
\title{setSlop generic}
\usage{
setSlop(object, treatment.only = TRUE)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}

\item{treatment.only}{Default \code{TRUE}, take only first periode of treatment where dose is not zero}
}
\value{
a \code{list} with slop of each model.id
}
\description{
Generic for setSlop method
}
\details{
If dose column is not present treatment.only will be considered FALSE
}
\examples{
data(pdxe)
slops <- setSlop(object = pdxe, treatment.only=FALSE)
}

