% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{setSlope<-}
\alias{setSlope<-}
\title{setSlope<- Generic}
\usage{
setSlope(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}
}
\value{
Updated \code{XevaSet}
}
\description{
Generic for setSlope replace method
}
\examples{
data(pdxe)
setSlope(pdxe) <- setSlope(object = pdxe)
}
