% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{setSlope}
\alias{setSlope}
\title{setSlope generic}
\usage{
setSlope(object, treatment.only = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}

\item{treatment.only}{Default \code{TRUE}, take only first periode of treatment where dose is not zero}
}
\value{
a \code{list} with slope of each model.id
}
\description{
Generic for setSlope method
}
\details{
If dose column is not present treatment.only will be considered FALSE
}
\examples{
data(pdxe)
slope <- setSlope(object = pdxe, treatment.only=FALSE)
}
