% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoPX_Class.R
\docType{methods}
\name{show,PharmacoPSet-method}
\alias{show,PharmacoPSet-method}
\title{A method to display object
for "show" setGeneric is already defined}
\usage{
\S4method{show}{PharmacoPSet}(object)
}
\description{
A method to display object
for "show" setGeneric is already defined
}

