% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoSet_Class.R
\docType{methods}
\name{show,XenoSet-method}
\alias{show,XenoSet-method}
\title{A method to display object
for "show" setGeneric is already defined}
\usage{
\S4method{show}{XenoSet}(object)
}
\description{
A method to display object
for "show" setGeneric is already defined
}

