% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_Xeva.R
\name{subsetXeva}
\alias{subsetXeva}
\title{Subset Xeva object.}
\usage{
subsetXeva(object, ids, id.name, keep.batch = TRUE)
}
\arguments{
\item{object}{The \code{XevaSet} object.}

\item{ids}{IDs to be selected for.}

\item{id.name}{Names of the IDs.}

\item{keep.batch}{Default \code{TRUE}. If \code{FALSE}, remove all other \code{model.ids} from the experiemt design that do not belong to selection.}
}
\value{
New Xeva object.
}
\description{
Subset Xeva object.
}
\examples{
data(brca)
print(brca)
df <- subsetXeva(brca, ids = c("X-1004", "X-1008", "X-1286"), id.name="patient.id", keep.batch=TRUE)
print(df)
}
