% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_MolData_Functions.R
\name{summarizeData}
\alias{summarizeData}
\alias{summarizeMolecularProfiles}
\title{Summarize molecular profiles and drug response data}
\usage{
summarizeData(
  object,
  drug,
  mDataType,
  tissue = NULL,
  sensitivity.measure = "all",
  sen.type = c("batch", "model"),
  unique.model = TRUE,
  batch = NULL,
  summarizedExp = FALSE
)

summarizeMolecularProfiles(...)
}
\arguments{
\item{object}{The \code{XevaSet}.}

\item{drug}{Name of the drug.}

\item{mDataType}{\code{character}, where one of the molecular data types is needed.}

\item{tissue}{Default \code{NULL} will return all tissue types.}

\item{sensitivity.measure}{Default \code{'all'} will return all sensitivity measures.}

\item{sen.type}{Type of sensitivity measure. Options are 'batch' (default) or 'model'}

\item{unique.model}{Default \code{TRUE} will return only one sequncing ID, in the case where one model ID maps to several sequencing IDs.}

\item{batch}{Name of the batch. Default \code{NULL}.}

\item{summarizedExp}{If \code{TRUE} will return SummarizedExperiment class object. Default \code{TRUE} will return ExpressionSet.}
}
\value{
An \code{ExpressionSet} where sample names are \code{model.id} and sensitivity measures will be presented in \code{pData}.
}
\description{
This function serves to get molecular profiles from a \code{XevaSet} object.
}
\details{
\itemize{
\item {If a sequencing sample belongs to multiple models, this function
will create a separate column for each model.}
\item {All models without molecular data will be removed from the output object.}
}
}
\examples{
data(brca)

pacRNA <- summarizeData(brca, drug="paclitaxel", mDataType="RNASeq",
                    tissue= "BRCA", sensitivity.measure="mRECIST")
print(pacRNA)

#to get all batch level response
pacRNA <- summarizeData(brca, drug="paclitaxel", mDataType="RNASeq",
                    tissue= "BRCA", sensitivity.measure="all")
print(pacRNA)

#to get all model level response
pacRNA <- summarizeData(brca, drug="paclitaxel", mDataType="RNASeq",
                    tissue= "BRCA", sensitivity.measure="all",
                    sen.type="model")
print(pacRNA)
}
