% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_response_Function.R
\name{summarizeResponse}
\alias{summarizeResponse}
\title{Summarize Response of PDXs}
\usage{
summarizeResponse(object, response.measure = "mRECIST_recomputed",
  group.by = "patient.id", summary.stat = c(";", "mean", "median"),
  tumor.type = NULL)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{response.measure}{\code{character} . Which response measure to use? Use the responseMeasures function to find out what measures are available for each Xeva set.}

\item{group.by}{default \code{patient.id}. How the models should be grouped togather. See details}

\item{summary.stat}{which summary method to use if multipal ids were found}
}
\value{
a \code{matrix} with rows as drug names, coulmn as \code{group.by} and each cell contains \code{response.measure} for the pair.
}
\description{
Summarize Response of PDXs.
}
\details{
There can be two types of response measure
\itemize{
\item{per model response : One response value for each Model, e.g. mRECIST_recomputed for each model}
\item{per batch response : One response value for each Batch, e.g. angle between treatment and control groups}
}
In case of \code{per model response} output columns will be \code{model.id} (or group.by).
For \code{per batch response} \code{group.by} value can be \code{"batch.name"} .
}
\examples{
data(pdxe)
pdxe_mR <- summarizeResponse(pdxe, response.measure = "mRECIST_recomputed", group.by="patient.id")
#to get only lung PDXE
pdxe_mR <- summarizeResponse(pdxe, response.measure = "mRECIST_recomputed",
                             group.by="patient.id", tumor.type="NSCLC")
}

