% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{symmetricSetDiff}
\alias{symmetricSetDiff}
\title{Symmetric set differences}
\usage{
symmetricSetDiff(a, b)
}
\arguments{
\item{a}{vector}

\item{b}{vector}
}
\value{
Returns a vector
}
\description{
\code{symmetricSetDiff} give symmetric set differences. Symmetric set difference (disjunctive union), of two sets is the set of elements which are not in their intersection.
}
\examples{
a <- c(1, 2, 3)
b <- c(2, 3, 4)
symmetricSetDiff(a, b)
}

