% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgi.R
\name{TGI}
\alias{TGI}
\title{tumor growth inhibition (TGI)
Computes the tumor growth inhibition (TGI) between two time-volume curves}
\usage{
TGI(contr.volume, treat.volume)
}
\arguments{
\item{contr.volume}{Volume vector for control}

\item{treat.volume}{Volume vector for treatment}
}
\value{
Returns batch response object
}
\description{
tumor growth inhibition (TGI)
Computes the tumor growth inhibition (TGI) between two time-volume curves
}
\examples{
contr.volume <- c(1.35, 6.57, 13.94, 20.39, 32.2, 39.26, 46.9, 53.91)
treat.volume <- c(0.4, 1.26, 2.59, 3.62, 5.77, 6.67, 7.47, 8.98, 9.29, 9.44)
TGI(contr.volume, treat.volume)

}
