% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{volume}
\alias{volume}
\title{compute PDX volume
\code{volume} computes PDXs tumor volume from tumor width and length}
\usage{
volume(w, l)
}
\arguments{
\item{w}{width or first measurement vector. See \code{Details}}

\item{l}{length or second measurement vector. See \code{Details}}
}
\value{
Returns a vector of volume
}
\description{
compute PDX volume
\code{volume} computes PDXs tumor volume from tumor width and length
}
\details{
This function applies the (l x w^2)/2 formula to compute volume. However
before computing volume it sorts data so that the smaller value is always taken as
width and  the bigger value as length.
}
\examples{
volume(w=1:5, l=6:10)
## smaller value is always considered as width.
volume(w=6:10, l=1:5)

}
