% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{angle}
\alias{angle}
\title{Computes angle
Compute angle for given models or batch}
\usage{
angle(object, model.id = NULL, batchName = NULL, expDig = NULL,
  treatment.only = TRUE, max.time = NULL, impute.value = TRUE,
  return.fit = FALSE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Computes angle
Compute angle for given models or batch
}
\examples{
data(pdxe)
angle(object=pdxe, model.id=c("X.007.BG98", "X.6047.uned"))
angle(object=pdxe, batchName="X-6047.paclitaxel")
}
