% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R,
%   R/slope_and_angle.R
\name{angle}
\alias{angle}
\alias{angle}
\title{Computes angle
Compute angle for given models or batch}
\usage{
angle(time1, volume1, time2 = NULL, volume2 = NULL, degree = TRUE)

angle(time1, volume1, time2 = NULL, volume2 = NULL, degree = TRUE)
}
\arguments{
\item{time1}{first vector of time}

\item{volume1}{first vector of volume}

\item{time2}{second vector of time}

\item{volume2}{second vector of volume}

\item{degree}{default \code{TRUE} will give angle in Degree and \code{FALSE} will return Radians}

\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id} for which slope is required}

\item{batchName}{batch name}

\item{expDig}{Experiment design list}

\item{treatment.only}{default \code{FALSE}. Given full data treatment.only=\code{TRUE} will plot data only during treatment}

\item{max.time}{maximum time point of the plot, default \code{NULL} will plot complete data}

\item{impute.value}{default \code{TRUE} will impute missing values}

\item{return.fit}{default \code{FALSE}. If \code{TRUE} will return \code{lm} fit also}
}
\value{
a \code{data.fram} with treatment, control and batch.name

returns angle and \code{slope} object
}
\description{
Computes angle
Compute angle for given models or batch

\code{angle} returns the angle between two volume data
}
\examples{
data(pdxe)
angle(object=pdxe, model.id=c("X.007.BG98", "X.6047.uned"))
angle(object=pdxe, batchName="X-6047.paclitaxel")
time  <- c(0, 3, 7, 11, 18, 22, 26, 30, 32, 35)
volume1<- time * tan(30*pi/180)
volume2<- time * tan(45*pi/180)
ang <- angle(time, volume1, time, volume2)
par(pty="s")
xylimit <- range(c(time, volume1, time, volume2))
plot(time, volume1, type = "b", xlim = xylimit, ylim = xylimit)
lines(time, volume2, type = "b")
abline(lm(volume1~time))
abline(lm(volume2~time))
}
