% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{batchInfo}
\alias{batchInfo}
\title{Get batch information}
\usage{
batchInfo(object, batch = NULL, model.id = NULL, model.id.type = c("any",
  "control", "treatment"))
}
\arguments{
\item{object}{xeva object}

\item{batch}{name of the batch. Default \code{NULL}}

\item{model.id}{model id for which need to be searched in the batches. Default \code{NULL}}

\item{model.id.type}{type of the model id in a batch. See details}
}
\value{
A \code{Vector} with batch names
}
\description{
Get batch information from a Xeva dataset. By default it will return the names of all the batch present in the data-set.
\code{model.id} is specified, will return all batch names conting that model.id
}
\details{
By default it will return the names of all the batch present in the
data-set. If batch specified it will retun the experiment design (control
and treatment model ids) of that perticular batch. If model.id is specified
it will return names of all the batches where this perticuler model.id is present.

For model.id.type the default value \code{'any'} will return all batch ids
where given model id is present in any arm (control or treatment) of the
batch. It can be restriceted to look only for treatment (or control) arm by
specifying the type.
}
\examples{
data(brca)
##to get all the batch names
batch.name <- batchInfo(brca)

##to get an specific batch
batch.design <- batchInfo(brca, batch=batch.name[1])

##to get all the batchs where a model.id is present
batchInfo(brca, model.id="X.6047.uned")

}
