% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{batchNames}
\alias{batchNames}
\title{Get batch names/ids}
\usage{
batchNames(object, model.id = NULL)
}
\arguments{
\item{object}{A \code{XevaSet}}

\item{model.id}{default \code{NULL}. if specified it will return batch names conting that model.id}
}
\value{
A \code{Vector} with batch names
}
\description{
Get batch names/ids from a Xeva dataset. If \code{model.id} is specified, will return all batch names conting that model.id
}
\examples{
data(brca)
batchNames(brca)
batchNames(brca, model.id="X.6047.uned")
}
