% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{calculateAngle}
\alias{calculateAngle}
\title{Calculate angle between control and treatment groups}
\usage{
calculateAngle(object, ExpDesign = NULL, var = "volume",
  treatment.only = TRUE, plot = TRUE, log.y = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{ExpDesign}{A list with batch.name, treatment and control}

\item{var}{Name of the variable, default \code{volume}}

\item{treatment.only}{default {TRUE}. If TRUE only treatment periode will be considered}

\item{plot}{default \code{TRUE} will return a list with element \code{angle} value and plot}

\item{log.y}{default \code{TRUE} will take log of the values before linear regression. After taking log all non-numeric (-Inf, Inf, NaN) values will be ignored}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a batch (control and treatment model ids)
it will return angle (in Degree) between
the linear fit of treatment and control group
}
\details{
If plot==TRUE, this will return a list with angle and plot. The plot is made using ggplot and can be ggplot customized for aesthetics. See example for details
}
\examples{
data(pdxe)
# creat a experiment desing
ExpDesign = list(batch.name="myBatch", treatment=c("X.015.BY19"), control=c("X.015.uned"))
angl = calculateAngle(object=pdxe, ExpDesign, var = "volume", treatment.only=TRUE, plot=TRUE)
#print angle
print(angl$angle)
#print plot
print(angl$myBatch$plot)
#print without legend
print(angl$myBatch$plot+ theme(legend.position='none'))
}

