% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_angle_functions.R
\name{computAngelFor1ExpDesign}
\alias{computAngelFor1ExpDesign}
\title{expDegI = list(batch.name = "myBatch", treatment = "X.015.BY19", control = "X.015.uned")}
\usage{
computAngelFor1ExpDesign(object, expDegI, var = "volume",
  treatment.only = TRUE, plot = FALSE)
}
\description{
expDegI = list(batch.name = "myBatch", treatment = "X.015.BY19", control = "X.015.uned")
}

