% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRECIST.R
\name{computemRECIST}
\alias{computemRECIST}
\title{Computes the mRECIST}
\usage{
computemRECIST(time, volume, min.time = 10, return.detail = FALSE)
}
\arguments{
\item{time}{Value of best response}

\item{volume}{Value of best average response}

\item{min.time}{minimum time after which tumore volume will be considered}

\item{return.detail}{default \code{FALSE}. If \code{TRUE} will return all intermediate values}
}
\value{
Returns the mRECIST
}
\description{
\code{computemRECIST} returns the mRECIST for given volume response
}
\examples{
time  <- c(0, 3, 7, 11, 18, 22, 26, 30, 32, 35)
volume<- c(250.8, 320.4, 402.3, 382.6, 384, 445.9, 460.2, 546.8, 554.3, 617.9)
computemRECIST(time, volume, min.time=10, return.detail=FALSE)
}
