% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XevaSet_Class.R
\name{createXevaSet}
\alias{createXevaSet}
\title{XevaSet constructor}
\usage{
createXevaSet(name, model = data.frame(), drug = data.frame(),
  experiment = data.frame(), expDesign = list(),
  modelSensitivity = data.frame(), batchSensitivity = data.frame(),
  molecularProfiles = list(), modToBiobaseMap = data.frame())
}
\arguments{
\item{name}{A \code{character} string detailing the name of the dataset.}

\item{model}{A \code{data.frame} containing the annotations for all the models used
in the experiment.}

\item{drug}{A \code{data.frame} containing the annotations for all the drugs
profiled in the dataset, across all data types.}

\item{experiment}{A \code{data.frame} containing all experiment information.}

\item{expDesign}{a list containing name of the batch, control and treatment model.id}

\item{modelSensitivity}{data.fram containg sensitivity for each model}

\item{batchSensitivity}{data.fram containg sensitivity for each batch}

\item{molecularProfiles}{A \code{list} of \code{ExpressionSet} objects containing
different molecular profiles.}

\item{modToBiobaseMap}{a data.fram containg model.id corresponding Biobase object id and name of the molecularProfiles}
}
\value{
Returns Xeva object
}
\description{
A constructor to create XevaSet. Only objects returned by this constructor
are expected to work with the XevaSet methods.
}
\examples{
NULL

}
