% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XenoSet_Class.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{drugInfo<- Generic
Generic for drugInfo replace method}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{The \code{XenoSet} to replace drug info in}

\item{value}{A \code{data.frame} with the new drug annotations}
}
\value{
Updated \code{XenoSet}
}
\description{
drugInfo<- Generic
Generic for drugInfo replace method
}
\examples{
data(pdxe)
drugInfo(pdxe) <- drugInfo(pdxe)
}

