% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_drug.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{set drug information
set drug information slot}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}

\item{value}{A \code{data.frame} with the new drug annotations}
}
\value{
updated \code{XevaSet}
}
\description{
set drug information
set drug information slot
}
\examples{
data(brca)
drugInfo(brca)<- drugInfo(brca)
}
