% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_drug.R
\name{drugInfo}
\alias{drugInfo}
\title{Get drug information
Get the drug information slot from a {XevaSet} object.}
\usage{
drugInfo(object)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug information from.}
}
\value{
A \code{data.frame} with the drug annotations.
}
\description{
Get drug information
Get the drug information slot from a {XevaSet} object.
}
\examples{
data(brca)
drugInfo(brca)
}
