% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_drug.R
\name{drugInfo}
\alias{drugInfo}
\title{get drug information
get drug information slot}
\usage{
drugInfo(object)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the drug annotations
}
\description{
get drug information
get drug information slot
}
\examples{
data(brca)
drugInfo(brca)
}
