% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PharmacoPX_Class.R
\name{drugInfo}
\alias{drugInfo}
\title{drugInfo Generic
Generic for drugInfo method}
\usage{
drugInfo(object)
}
\arguments{
\item{object}{The \code{PharmacoPSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the drug annotations
}
\description{
drugInfo Generic
Generic for drugInfo method
}
\examples{
data(pdxe)
drugInfo(pdxe)
}

