% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_Sensitivity_Signature.R
\name{drugSensitivitySig}
\alias{drugSensitivitySig}
\title{drugSensitivitySig}
\usage{
drugSensitivitySig(object, drug, mDataType = NULL, molData = NULL,
<<<<<<< HEAD
  features = NULL, model.ids = NULL, model2bidMap = NULL,
  sensitivity.measure = "slope", fit = c("lm", "maxCor", "gam"),
  standardize = c("SD", "rescale", "none"), nthread = 1,
  tumor.type = NULL, verbose = TRUE)
=======
  features = NULL, model.ids = NULL, sensitivity.measure = "slope",
  fit = c("lm", "maxCor", "gam"), standardize = c("SD", "rescale", "none"),
  nthread = 1, tumor.type = NULL, verbose = TRUE)
>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{drug}{Name of the drug}

\item{mDataType}{molecular data type}

\item{molData}{External data matrix. Rows as features and columns as samples}

<<<<<<< HEAD
\item{features}{which  molecular data fetures to use. Default \code{NULL} will use all fetures}

\item{model.ids}{which model.id to use from the dataset. Default \code{NULL} will use all model.id}

\item{model2bidMap}{a datafram with model.id and biobase.id. Default \code{NULL} will use internal mapping}
=======
\item{features}{Which fetures to use from Biobase object. Default \code{NULL} will use all fetures.}
>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4

\item{sensitivity.measure}{Name of the sensitivity measure}

\item{fit}{Default \code{lm}. Name of the model to be fitted. Options are "lm", "maxCor", "gam"}

\item{type}{Tissue type. Default is NULL which will use \code{'tumor.type'} from \code{object}}
}
\value{
A datafram with fetures and values
}
\description{
Given a Xeva object, and drug name it will return sensitivity value for all the genes/fetures
<<<<<<< HEAD
}
\details{
A matrix of values can be directly passed to molData. \code{fit} can be "lm", "maxCor" or "gam".
In case where a model.id map to multipal biobase.id the first biobase.id in the datafram will be used.
=======

A matrix of values can be directly passed to molData. \code{fit} can be "lm", "maxCor" or "gam"
}
\details{
Get sensitivity signatures for a drug
>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
}
\examples{
data(pdxe)
## select BRCA samples
mid <- modelInfo(pdxe)[modelInfo(pdxe)$tumor.type=="BRCA", ]
drugSensitivitySig(object=pdxe, drug=c("paclitaxel","tamoxifen"),
                   mDataType="RNASeq", features=1:5,
                   model.ids = mid$model.id,
                   sensitivity.measure="slope", fit = "lm")
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
