% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_Sensitivity_Signature.R
\name{drugSensitivitySig}
\alias{drugSensitivitySig}
\title{drugSensitivitySig}
\usage{
drugSensitivitySig(object, mDataType, drug, features = NULL,
  sensitivity.measure = "slop", standardize = c("SD", "rescale", "none"),
  nthread = 1, verbose = TRUE, ...)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{drug}{Name of the drug}

\item{features}{Which fetures to use from Biobase object. Default \code{NULL} will use all fetures.}
}
\value{
A datafram with fetures and values
}
\description{
Given a Xeva object, and drug name it will return sensitivity value for all the genes/fetures
}
\details{
Get sensitivity signatures for a drug
}
\examples{
data(cm.pdxe)
drugSensitivitySig(object=cm.pdxe, mDataType="RNASeq", drug="binimetinib", features=1:5,
sensitivity.measure="slop")

}

