% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_Sensitivity_Signature.R
\name{drugSensitivitySig}
\alias{drugSensitivitySig}
\title{drugSensitivitySig}
\usage{
drugSensitivitySig(object, drug, mDataType = NULL, molData = NULL,
  features = NULL, sensitivity.measure = "slope", fit = c("lm", "maxCor",
  "gam"), standardize = c("SD", "rescale", "none"), nthread = 1,
  type = NULL, verbose = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{drug}{Name of the drug}

\item{mDataType}{molecular data type}

\item{molData}{External data matrix. Rows as features and columns as samples}

\item{features}{Which fetures to use from Biobase object. Default \code{NULL} will use all fetures.}

\item{sensitivity.measure}{Name of the sensitivity measure}

\item{fit}{Default \code{lm}. Name of the model to be fitted. Options are "lm", "maxCor", "gam"}

\item{type}{Tissue type. Default is NULL which will use \code{'tumor.type'} from \code{object}}
}
\value{
A datafram with fetures and values
}
\description{
Given a Xeva object, and drug name it will return sensitivity value for all the genes/fetures

A matrix of values can be directly passed to molData. \code{fit} can be "lm", "maxCor" or "gam"
}
\details{
Get sensitivity signatures for a drug
}
\examples{
data(pdac.pdxe)
drugSensitivitySig(object=pdac.pdxe, drug="binimetinib", mDataType="RNASeq", features=1:5,
sensitivity.measure="slope", fit = "lm")

}

