% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_Sensitivity_Signature.R
\name{drugSensitivitySig}
\alias{drugSensitivitySig}
\title{drugSensitivitySig}
\usage{
drugSensitivitySig(object, mDataType, drug, features = NULL,
  sensitivity.measure = "slop", fit = c("lm", "maxCor", "gam"),
  standardize = c("SD", "rescale", "none"), nthread = 1, type = NULL,
  verbose = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{mDataType}{molecular data type}

\item{drug}{Name of the drug}

\item{features}{Which fetures to use from Biobase object. Default \code{NULL} will use all fetures.}

\item{sensitivity.measure}{Name of the sensitivity measure}

\item{fit}{Default \code{lm}. Name of the model to be fitted. Options are "lm", "maxCor", "gam"}

\item{type}{Tissue type. Default is NULL which will use \code{'tumor.type'} from \code{object}}
}
\value{
A datafram with fetures and values
}
\description{
Given a Xeva object, and drug name it will return sensitivity value for all the genes/fetures
}
\details{
Get sensitivity signatures for a drug
}
\examples{
data(cm.pdxe)
drugSensitivitySig(object=cm.pdxe, mDataType="RNASeq", drug="binimetinib", features=1:5,
sensitivity.measure="slop", fit = c("lm", "maxCor", "gam"))

}

