% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{drugWaterfall}
\alias{drugWaterfall}
\title{Plot drug response waterfall}
\usage{
drugWaterfall(object, drug, value, model.ids = NULL, model.col = "#cc4c02",
  sortByType = FALSE, title = "Waterfall plot", yname = "Response",
  tumor.type = NULL)
}
\arguments{
\item{object}{the \code{XevaSet}}

\item{drug}{name of the drug}

\item{value}{which value should be ploted}

\item{model.col}{color for model.ids. A data.frame with two colomns "color" and "type" can be provided}

\item{model.id}{which model.ids to plot. Default is \code{NULL} will plot all models}
}
\description{
Plot drug response waterfall
}
\examples{
data(pdxe)
drugWaterfall(pdxe, drug="binimetinib", value="best.avg.response_published",
              title="Binimetinib", yname = "Change in tumor volume (\%)",
              tumor.type = "CRC")
}
