% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{expDesign}
\alias{expDesign}
\title{Given a batch.name get batch}
\usage{
expDesign(object, batch.name)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{object}{The \code{batch.name}}
}
\value{
A \code{Vector} with all batch.name
}
\description{
Given a batch.name get batch from a Xeva dataset
}
\examples{
data(pdxe)
expDesign(pdxe, batch.name = "X-6047.paclitaxel")
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
