% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getBatchName}
\alias{getBatchName}
\title{Get batch.name for a given model.id}
\usage{
getBatchName(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id} for which batch name required}
}
\value{
a \code{vector} with all batch names
}
\description{
Get batch.name for a given model.id.
If no batch.name found it will return NULL
}
\examples{
data(pdxe)
# extract batch.name for a given model.id
getBatchName(object=pdxe, model.id="X.1655.uned")
getBatchName(object=pdxe, model.id="X.010.fiab")
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
