% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getExpDesignDF}
\alias{getExpDesignDF}
\title{Given a model.id it will return a data.fram of experiemt design
with columns as "treatment", "control", "batch.name"}
\usage{
getExpDesignDF(object, model.id)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a model.id it will return a data.fram of experiemt design
with columns as "treatment", "control", "batch.name"
}
\examples{
data(pdxe)
# This will give a data.fram with columns as "treatment", "control", "batch.name"
getExpDesignDF(object=pdxe, model.id="X.1655.LE11.biib")
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
