% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{getExperiment}
\alias{getExperiment}
\title{getExperiment Generic
Generic for getExperiment method}
\usage{
getExperiment(object, experiment.id)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{list} with the all experiment designs
}
\description{
getExperiment Generic
Generic for getExperiment method
}
\examples{
data(pdxe)
getExperiment(pdxe, experiment.id="X.1004.pael.paclitaxel")
getExperiment(pdxe, experiment.id=c("X.1286.pael.paclitaxel", "X.1298.pael.paclitaxel") )
}

