% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{getExperiment}
\alias{getExperiment}
\title{For a given  model.id, it will return a data.fram
containing all data stored in experiment slot}
\usage{
getExperiment(object, model.id = NULL, batch.name = NULL,
  treatment.only = FALSE)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{model.id}{The \code{model.id} for which data is required}

\item{treatment.only}{Default \code{FALSE}. If TRUE give data only for non-zero dose periode (if dose data avalible)}
}
\value{
a \code{data.fram} will all the the values stored in experiment slot
}
\description{
For a given  model.id, it will return a data.fram
containing all data stored in experiment slot
}
\examples{
data(pdxe)
getExperiment(pdxe, model.id="X.1004.pael")
}

