% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_batch_data.R
\name{getExperiment}
\alias{getExperiment}
\title{For a given  model.id, it will return a data.fram
containing all data stored in experiment slot}
\usage{
getExperiment(object, model.id = NULL, batch = NULL, patient.id = NULL,
  drug = NULL, control.name = NULL, treatment.only = FALSE,
  max.time = NULL, vol.normal = FALSE, return.list = FALSE,
  impute.value = FALSE, concurrent.time = FALSE)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{model.id}{The \code{model.id} for which data is required, multipal allowed}

\item{batch}{batch name from the Xeva set or experiment design}

\item{treatment.only}{Default \code{FALSE}. If TRUE give data only for non-zero dose periode (if dose data avalible)}

\item{max.time}{maximum time for data}

\item{vol.normal}{default \code{TRUE} will use}

\item{return.list}{default \code{FALSE} will return a datafram}

\item{impute.value}{default \code{FALSE}. If TRUE will impute the values}

\item{concurrent.time}{default \code{FALSE}. If TRUE will cut the batch data such that control and treatment will end at same time point}
}
\value{
a \code{data.fram} will all the the values stored in experiment slot
}
\description{
For a given  model.id, it will return a data.fram
containing all data stored in experiment slot
}
\examples{
data(brca)
getExperiment(brca, model.id="X.6047.uned", treatment.only=TRUE)
getExperiment(brca, model.id=c("X.6047.uned", "X.6047.pael"), treatment.only=TRUE)
getExperiment(brca, batch="X-6047.paclitaxel", treatment.only=TRUE)
ed <- list(batch.name="myBatch", treatment=c("X.6047.LJ16","X.6047.LJ16.trab"),
             control=c("X.6047.uned"))

getExperiment(brca, batch=ed)

}
