% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_expressionData_Functions.R
\name{getMolecularProfiles}
\alias{getMolecularProfiles}
\title{Get Molecular Profiles}
\usage{
getMolecularProfiles(object, data.type)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{data.type}{\code{character}, which one of the molecular data types is needed}
}
\value{
a \code{ExpressionSet} where sample names are \code{biobase.id} of model
}
\description{
Get Molecular Profiles
}
\examples{
data(pdxe)
pdxe_RNA <- getMolecularProfiles(pdxe, data.type="RNASeq")
}

