% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_expDesign.R
\name{getTimeVarData}
\alias{getTimeVarData}
\title{Given a model.id it will return a data.fram of experiemt design}
\usage{
getTimeVarData(object, ExpDesign, var, collapse)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{ExpDesign}{A list with batch.name, treatment and control}
}
\value{
a \code{data.fram} with treatment, control and batch.name
}
\description{
Given a model.id it will return a data.fram of experiemt design
}
\examples{
data(pdxe)
# extract controls for a given model.id
ExpDesign = list(batch.name="myBatch", treatment=c("X.010.BG98"), control=c("X.010.uned"))
getTimeVarData(object=pdxe, ExpDesign, var = "volume", collapse=TRUE)
}

