% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm.R
\name{lmm}
\alias{lmm}
\title{linear mixed model}
\usage{
lmm(data, log.volume)
}
\arguments{
\item{data}{a data.frame containing a batch data}

\item{log.volume}{FALSE if volume is raw, TRUE if volume is in log}
}
\value{
Returns a fit object
}
\description{
Comput the linear mixed model (lmm) statistics for a PDX batch
}
\details{
The input data.frame (data) must contain these columns: 'model.id', 'volume', 'time', 'exp.type'.
Lower value from lmm function indicates better response.
}
\examples{
data(repdx)
data <- getExperiment(repdx, batch = "P1")$model
lmm(data, log.volume=FALSE)

data3 <- getExperiment(repdx, batch = "P3")$model
lmm(data3, log.volume=FALSE)

}
