% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{mapModelSlotIds}
\alias{mapModelSlotIds}
\title{Map ids of model slot}
\usage{
mapModelSlotIds(object, id, id.name, map.to = "all", unique = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{id}{The \code{id}}

\item{id.name}{The \code{id} name}

\item{map.to}{The name of the mapped id. Default \code{all}}

\item{unique}{Default \code{TRUE}. If unique=FALSE output will be mapped to input}
}
\value{
a \code{data.fram} with id and mapped id
}
\description{
Map one id type to another in model slot.
For example map a model.id to patient.id
}
\examples{
data(pdxe)
mapModelSlotIds(object=pdxe, id="X-007", id.name="patient.id", map.to="model.id")
##map batch ids
mapModelSlotIds(pdxe, id= "X-011.INC280", id.name = "batch.name", map.to = "tumor.type")
}
