% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{modelInfo<-}
\alias{modelInfo<-}
\title{modelInfo<- Generic
Generic for modelInfo replace method}
\usage{
modelInfo(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}

\item{value}{A \code{data.frame} with the new model annotations}
}
\value{
Updated \code{XevaSet}
}
\description{
modelInfo<- Generic
Generic for modelInfo replace method
}
\examples{
data(pdxe)
modelInfo(pdxe) <- modelInfo(pdxe)
}
