% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{modelInfo}
\alias{modelInfo}
\title{modelInfo Generic
Generic for modelInfo method}
\usage{
<<<<<<< HEAD
modelInfo(object, mDataType = NULL)
=======
modelInfo(object)
>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the model annotations
}
\description{
modelInfo Generic
Generic for modelInfo method
}
\examples{
data(pdxe)
<<<<<<< HEAD
mid <- modelInfo(pdxe)
head(mid)
}
=======
modelInfo(pdxe)
}

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
