% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{modelInfo}
\alias{modelInfo}
\title{modelInfo Generic
Generic for modelInfo method}
\usage{
modelInfo(object, mDataType = NULL)
}
\arguments{
\item{object}{The \code{XevaSet} to retrieve drug info from}
}
\value{
a \code{data.frame} with the model annotations
}
\description{
modelInfo Generic
Generic for modelInfo method
}
\examples{
data(brca)
mid <- modelInfo(brca)
head(mid)
}
