% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_model.R
\name{modelInfo}
\alias{modelInfo}
\alias{modelInfo,XevaSet-method}
\title{modelInfo Generic
Generic for modelInfo method}
\usage{
modelInfo(object, mDataType = NULL)

\S4method{modelInfo}{XevaSet}(object, mDataType = NULL)
}
\arguments{
\item{object}{Xeva object}

\item{mDataType}{Molecular data type.}
}
\value{
A \code{data.frame} with the model annotations.
}
\description{
modelInfo Generic
Generic for modelInfo method
}
\examples{
data(brca)
mid <- modelInfo(brca)
head(mid)

}
