% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/important_functions.R
\name{pasteColTogather}
\alias{pasteColTogather}
\title{paste a data.frame columns togather while removing NA}
\usage{
pasteColTogather(df, collapse = " + ")
}
\arguments{
\item{df}{A data.frame}

\item{collapse}{Collapse string default " + "}
}
\value{
Returns an vector of strings where column values paste togather
}
\description{
\code{pasteColTogather} paste a data.frame columns togather while removing NA
}
\examples{
df = data.frame(x= 1:6, y = c("A", NA, "B", NA, NA, "C"))
pasteColTogather(df, collapse = " + ")
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
