% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotBatch}
\alias{plotBatch}
\title{data(lpdx)
plotBatch(lpdx, "PHLC153_P6", treatment.only=FALSE, log.y=TRUE)}
\usage{
plotBatch(object, batchName = NULL, expDig = NULL, treatment.only = FALSE,
  control.col = "#6baed6", treatment.col = "#fc8d59", title = "",
  xlab = "Time", ylab = "Volume", log.y = FALSE, drgName = NULL,
  SE.plot = c("all", "none", "errorbar", "ribbon"), aspect.ratio = c(1,
  NULL), minor.line.size = 0.5, major.line.size = 0.7, max.time = NULL,
  vol.normal = FALSE)
}
\description{
data(lpdx)
plotBatch(lpdx, "PHLC153_P6", treatment.only=FALSE, log.y=TRUE)
}
