% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotBatch}
\alias{plotBatch}
\title{Plot batch data}
\usage{
plotBatch(object, batch = NULL, patient.id = NULL, drug = NULL,
  control.name = NULL, max.time = NULL, treatment.only = FALSE,
  vol.normal = FALSE, impute.value = TRUE, concurrent.time = FALSE,
  control.col = "#6baed6", treatment.col = "#fc8d59", title = "",
  xlab = "Time", ylab = "Volume", log.y = FALSE, drug.name = NULL,
  SE.plot = c("all", "none", "errorbar", "ribbon"), aspect.ratio = c(1,
  NULL), minor.line.size = 0.5, major.line.size = 0.7)
}
\arguments{
\item{object}{Xeva object}

\item{batch}{batch name or experiment design list}

\item{max.time}{maximum time point of the plot, default \code{NULL} will plot complete data}

\item{treatment.only}{default \code{FALSE}. Given full data treatment.only=\code{TRUE} will plot data only during treatment}

\item{vol.normal}{default \code{FALSE} . If TRUE volume will ne normalised}

\item{impute.value}{default \code{TRUE}, will impute values where missing}

\item{control.col}{color for control plots}

\item{treatment.col}{color for treatment plots}

\item{title}{title of the plot}

\item{xlab}{title of x axis}

\item{ylab}{title of y axis}

\item{log.y}{default \code{FALSE}, if \code{TRUE} y axis will be in log}

\item{drug.name}{default \code{NULL} will extract drug name from data}

\item{SE.plot}{plot type. Default \code{"all"} will plot all plots and average curves. Possible values are \code{"all"}, \code{"none"}, \code{"errorbar"}, \code{"ribbon"}}

\item{aspect.ratio}{default \code{1} will create equeal width and height plot}

\item{minor.line.size}{line size for minor lines default \code{0.5}}

\item{major.line.size}{line size for major lines default \code{0.7}}
}
\value{
A ggplot2 plot with control and treatment
}
\description{
Plot data for a batch id or experiment design
}
\examples{
data(brca)
plotBatch(brca, batch="X-1004.BGJ398", vol.normal=TRUE)
expDesign <- list(batch.name="myBatch", treatment=c("X.6047.LJ16","X.6047.LJ16.trab"),
             control=c("X.6047.uned"))
plotBatch(brca, batch=expDesign, vol.normal=T)
plotBatch(brca, batch=expDesign, vol.normal=F, SE.plot = "errorbar")
}
