% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotModelErrorBar}
\alias{plotModelErrorBar}
\title{##df = readRDS("DATA-raw/toPlot_DF.Rda")
##plotModelErrorBar(df)}
\usage{
plotModelErrorBar(dfp, control.col = "#6baed6", treatment.col = "#fc8d59",
  title = "", xlab = "Time", ylab = "Volume", log.y = FALSE,
  drgName = "", SE.plot = c("all", "none", "errorbar", "ribbon"),
  modelLyt = "dotted", aspect.ratio = c(1, NULL), minor.line.size = 0.5,
  major.line.size = 0.7)
}
\description{
##df = readRDS("DATA-raw/toPlot_DF.Rda")
##plotModelErrorBar(df)
}
