% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRECIST.R
\name{plotmRECIST}
\alias{plotmRECIST}
\title{To plot mRECIST values}
\usage{
plotmRECIST(mat, control.name = NA, control.col = "#238b45",
  drug.col = "black", colPalette = NULL, name = "Drug & Models",
  sort = TRUE, row_fontsize = 12, col_fontsize = 12, draw_plot = TRUE)
}
\arguments{
\item{mat}{The mRECIST matrix where rows are drugs and columns are patient}

\item{control.name}{name of the control}

\item{control.col}{color of the control}

\item{drug.col}{color of the drug names}

\item{colPalette}{color palette for mRECIST values}

\item{name}{title of the plot}

\item{sort}{if matrix should be sorted before ploting}

\item{row_fontsize}{size of the row name font}

\item{col_fontsize}{size of the column name font}

\item{draw_plot}{default TRUE will plot the figure. If FALSE will return an object}
}
\value{
plot
}
\description{
\code{plotmRECIST} plots the mRECIST matrix obtained from \code{summarizeResponse}
}
\examples{
data(brca)
## select lung cancer pdxe data
brca.mr <- summarizeResponse(brca, response.measure = "mRECIST", group.by="patient.id")
plotmRECIST(brca.mr, control.name = "untreated")
}
