% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRECIST.R
\name{plotmRECIST}
\alias{plotmRECIST}
\title{Plot mRECIST for models and drugs
\code{plot.mRECIST} plots the mRECIST}
\usage{
plotmRECIST(df, groupBy = "biobase.id", control.name = "untreated")
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
plot
}
\description{
Plot mRECIST for models and drugs
\code{plot.mRECIST} plots the mRECIST
}
\examples{
data(pdxe)
df = getmRECIST(pdxe)
df = df[1:500,]
plotmRECIST(df,groupBy = "biobase.id", control.name = "untreated")
}

