% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRECIST.R
\name{plotmRECIST}
\alias{plotmRECIST}
\title{To plot mRECIST values}
\usage{
plotmRECIST(mat, control.name = NA, control.col = "#238b45",
  drug.col = "black", colPalette = NULL, name = "Drug & Models",
  sort = TRUE)
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
plot
}
\description{
\code{plotmRECIST} plots the mRECIST matrix obtained from \code{summarizeResponse}
}
\examples{
data(pdxe)
## select lung cancer pdxe data
pdxe.lung <- summarizeResponse(pdxe, response.measure = "mRECIST",
                               group.by="patient.id", tumor.type="NSCLC")
plotmRECIST(pdxe.lung, control.name = "untreated")
}
