% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mRECIST.R
\name{plotmRECIST}
\alias{plotmRECIST}
\title{Plot mRECIST for models and drugs
\code{plot.mRECIST} plots the mRECIST}
\usage{
plotmRECIST(mat, control.name = NA, control.col = "green",
  drug.col = "black", colPalette = list(CR = "#377eb8", PR = "#4daf4a", SD =
  "#fec44f", PD = "#e41a1c"), name = "Drug & Models")
}
\arguments{
\item{object}{The \code{Xeva} dataset}

\item{model.id}{The \code{model.id}}
}
\value{
plot
}
\description{
Plot mRECIST for models and drugs
\code{plot.mRECIST} plots the mRECIST
}
\examples{
data(pdxe)
## select lung cancer pdxe data
pdxe.lung <- summarizeResponse(pdxe, response.measure = "mRECIST_recomputed",
                               group.by="patient.id", tumor.type="NSCLC")
plotmRECIST(pdxe.lung, control.name = "untreated")
}

