% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrices.R
\name{print.modelResponse}
\alias{print.modelResponse}
\title{Print the model response}
\usage{
\method{print}{modelResponse}(x, ...)
}
\arguments{
\item{x}{modelResponse object}

\item{...}{Other arguments}
}
\value{
prints the modelResponse
}
\description{
Print the model response
}
