% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_slot_experiment.R
\name{print.XevaSet}
\alias{print.XevaSet}
\title{print Xeva object}
\usage{
\method{print}{XevaSet}(object, id = NULL)
}
\arguments{
\item{object}{\code{Xeva} object}

\item{id}{default \code{NULL}, id can be \code{model.id} or \code{batch.name}}
}
\value{
Prints object, model or batch information.
}
\description{
\code{print} displays Xeva object information or model or batch information
}
\examples{
data(brca)
print(brca)

# to print a model
print(brca, id = "X.1004.BG98")

# to print a batch
print(brca, id = "X-1004.BGJ398")
}
\keyword{internal}
