% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUC.R
\name{setAAC<-}
\alias{setAAC<-}
\title{setAAC<- Generic}
\usage{
setAAC(object) <- value
}
\arguments{
\item{object}{The \code{XevaSet} to replace drug info in}
}
\value{
Updated \code{XevaSet}
}
\description{
Generic for setAAC replace method
}
\examples{
data(cm.pdxe)
setAAC(cm.pdxe) <- setAAC(object = cm.pdxe)
}
<<<<<<< HEAD
=======

>>>>>>> 9f9947748d00443b9546698266dd7eb78c636ce4
