% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_matrixs.R
\name{setResponse}
\alias{setResponse}
\title{\code{setResponse} sets response of an Xeva object}
\usage{
setResponse(object, res.measure = c("mRECIST", "slope", "AUC", "angle",
  "abc"), min.time = 10, treatment.only = TRUE, max.time = NULL,
  vol.normal = TRUE, impute.value = TRUE, concurrent.time = TRUE,
  verbose = TRUE)
}
\arguments{
\item{object}{Xeva object}

\item{res.measure}{response measure, multipal measure allowed}

\item{min.time}{default \strong{10} days. Used for \emph{mRECIST} computation}

\item{treatment.only}{Default \code{FALSE}. If TRUE give data only for non-zero dose periode (if dose data avalible)}

\item{max.time}{maximum time for data}

\item{vol.normal}{default \code{TRUE} will use}

\item{impute.value}{default \code{FALSE}. If TRUE will impute the values}

\item{concurrent.time}{default \code{FALSE}. If TRUE will cut the batch data such that control and treatment will end at same time point}

\item{verbose}{default \code{TRUE} will print infromation}
}
\value{
returns updated Xeva object
}
\description{
\code{setResponse} sets response of an Xeva object
}
\examples{
data(brca)
brca  <- setResponse(brca, res.measure = c("mRECIST"))
}
