% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XevaSet_Class.R
\docType{methods}
\name{show,XevaSet-method}
\alias{show,XevaSet-method}
\title{A method to display object
for "show" setGeneric is already defined}
\usage{
\S4method{show}{XevaSet}(object)
}
\description{
A method to display object
for "show" setGeneric is already defined
}
