% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_expressionData_Functions.R
\name{summarizeMolecularProfiles}
\alias{summarizeMolecularProfiles}
\title{summarizeMolecularProfiles}
\usage{
summarizeMolecularProfiles(object, drug, mDataType, tumor.type = NULL,
  sensitivity.measure = NULL)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{drug}{Name of the drug}

\item{mDataType}{\code{character}, which one of the molecular data types is needed}

\item{tumor.type}{default \code{NULL} will return all across all tumor.type}

\item{sensitivity.measure}{default \code{NULL} will return all sensitivity measure}
}
\value{
A \code{ExpressionSet} where sample names are model.id and sensitivity measure will be present in pData
}
\description{
summarizeMolecularProfiles
}
\details{
\itemize{
\item {If a sequencing sample belong to multipal models, summarizeMolecularProfiles
will creat saperate column for each model. }
\item {All the models without the moleculer data will be removed from the output expression set.}
}
}
\examples{
data(pdxe)
pdxe_RNA <- summarizeMolecularProfiles(pdxe, drug="paclitaxel", mDataType="RNASeq",
                                       tumor.type= "BRCA", sensitivity.measure="mRECIST")

}
