% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_expressionData_Functions.R
\name{summarizeMolecularProfiles}
\alias{summarizeMolecularProfiles}
\title{summarizeMolecularProfiles}
\usage{
summarizeMolecularProfiles(object, drug, mDataType, tissue = NULL,
  sensitivity.measure = NULL, unique.model = TRUE, batch = NULL)
}
\arguments{
\item{object}{The \code{XevaSet}}

\item{drug}{Name of the drug}

\item{mDataType}{\code{character}, which one of the molecular data types is needed}

\item{tissue}{default \code{NULL} will return all across all tissue}

\item{sensitivity.measure}{default \code{NULL} will return all sensitivity measure}

\item{unique.model}{default TRUE will return only one sequncing id, in case where one model id mapes to several sequencing ids}
}
\value{
A \code{ExpressionSet} where sample names are model.id and sensitivity measure will be present in pData
}
\description{
summarizeMolecularProfiles
}
\details{
\itemize{
\item {If a sequencing sample belong to multipal models, summarizeMolecularProfiles
will creat saperate column for each model. }
\item {All the models without the moleculer data will be removed from the output expression set.}
}
}
\examples{
data(brca)
pacRNA <- summarizeMolecularProfiles(brca, drug="paclitaxel", mDataType="RNASeq",
                                     tissue= "BRCA", sensitivity.measure="mRECIST")
print(pacRNA)
}
