% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall}
\alias{waterfall}
\title{waterfall plot
creates waterfall plot for a given drug}
\usage{
waterfall(object, drug, res.measure, group.by = NULL, tissue = NULL,
  model.id = NULL, model.type = NULL, type.color = "#cc4c02",
  legend.name = NULL, yname = NULL, title = NULL, sort = TRUE)
}
\arguments{
\item{object}{the \code{XevaSet}}

\item{drug}{name of the drug}

\item{res.measure}{PDX model response measure}

\item{group.by}{group response data}

\item{tissue}{tissue}

\item{model.id}{which model.id to plot. Default is \code{NULL} will plot all models}

\item{model.type}{type of model such as mutated or wild type}

\item{type.color}{a list with colors used for each type}

\item{legend.name}{name of the legend}

\item{yname}{name for y axis}

\item{title}{title of the plot}

\item{sort}{default TRUE will sort the data}
}
\description{
waterfall plot
creates waterfall plot for a given drug
}
\examples{
data(brca)
waterfall(brca, drug="binimetinib", res.measure="best.avg.response_published")
## example with model.type where we color the models by TP53 mutation type
mut <- summarizeMolecularProfiles(brca,drug = "binimetinib", mDataType="mutation")
model.type <- Biobase::exprs(mut)["TP53", ]
waterfall(brca, drug="binimetinib", res.measure="best.avg.response_published",
          tissue="BRCA", model.id=names(model.type), model.type= model.type)

}
