% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterfall_plot.R
\name{waterfall}
\alias{waterfall}
\title{waterfall plot
Creates waterfall plot for a given drug.}
\usage{
waterfall(
  object,
  res.measure = NULL,
  drug = NULL,
  group.by = NULL,
  summary.stat = c(";", "mean", "median"),
  x = NULL,
  y = NULL,
  type = NULL,
  type.color = NULL,
  na.value = "#878787",
  yname = NULL,
  title = NULL,
  sort = TRUE
)
}
\arguments{
\item{object}{The \code{XevaSet} object or a data.frame. See \code{Details}.}

\item{res.measure}{PDX model drug response measure}

\item{drug}{Name of the drug}

\item{group.by}{Group drug response data}

\item{summary.stat}{How to summarize multiple values. Options are ";", "mean" or "median". See \code{Details}.}

\item{x, y}{If object is data.frame, x and y indicates column names of x and y axis}

\item{type}{Type for each bar in waterfall (such as mutated or wild type). See Details.}

\item{type.color}{A color vector for type}

\item{na.value}{Color for NA values. Default "#878787"}

\item{yname}{Name for the y-axis}

\item{title}{Title of the plot}

\item{sort}{Default \code{TRUE} will sort the data}
}
\value{
waterfall plot in ggplot2
}
\description{
waterfall plot
Creates waterfall plot for a given drug.
}
\details{
The function waterfall can plot from a XevaSet or from a data.frame . 
If a data.frame is specified, x and y parameters must also be specified.
}
\examples{
data(brca)
##plot best.average.response for all models tested with binimetinib
waterfall(brca, drug="binimetinib", res.measure="best.average.response")

##plot same by taking mean of multiple models of each patient
waterfall(brca, drug="binimetinib", res.measure="best.average.response", 
          group.by = "patient.id", summary.stat = "mean")

## plot by specifing color by mutation type
## extract mutation information
mut <- summarizeData(brca,drug = "binimetinib", mDataType="mutation")
model.type <- Biobase::exprs(mut)["TP53", ]
## extract data.frame of response
df <- summarizeResponse(brca, response.measure = "best.average.response", return.type = "data.frame")
df <- df[df$drug == "binimetinib", ]
## add values to data.frame 
df$TP53 <- model.type[df$model.id]
## now plot the data 
waterfall(df, x="model.id", y="best.average.response", type="TP53")

}
